requirements.txt
flask
deepface
ultralytics
torch
torchvision
opencv-python
pillow
